Split event into services and stream
